+++ dyninstAPI/src/multiTramp.C
@@ -1564,7 +1564,7 @@ Address multiTramp::instToUninstAddr(Address addr) {
     }
 
     // Ran out of iterator... 
-    fprintf(stderr, "Checking %d deleted objects\n", deletedObjs.size());
+    fprintf(stderr, "Checking %ld deleted objects\n", (long) deletedObjs.size());
     for (unsigned i = 0; i < deletedObjs.size(); i++) {
         obj = deletedObjs[i];
         relocatedInstruction *insn = dynamic_cast<relocatedInstruction *>(obj);
